Add converter plugin helpers
