+++ resim/ros2/BUILD
@@ -244,6 +244,7 @@ cc_binary(
     linkshared = True,
     visibility = ["//resim/ros2:__subpackages__"],
     deps = [
+        ":converter_plugin_helpers",
         ":converter_plugin_types",
         ":detection_from_ros2",
         ":header_from_ros2",
@@ -253,6 +254,15 @@ cc_binary(
         ":pose_from_ros2",
         ":time_from_ros2",
         ":transform_from_ros2",
+        "//resim/assert",
+    ],
+)
+
+cc_library(
+    name = "converter_plugin_helpers",
+    hdrs = ["converter_plugin_helpers.hh"],
+    deps = [
+        ":converter_plugin_types",
         "//resim/assert",
         "//resim/utils/proto:dependency_file_descriptor_set",
         "@ros2_rclcpp//:rclcpp",
